+++ test/tests/error-handling-dev.test.js
@@ -33,7 +33,7 @@ describe('graphql - error handling in development', () => {
         {
           message: 'Value is required',
           extensions: {
-            code: '400',
+            code: 'ASSERT_NOT_NULL',
             target: 'notEmptyI',
             stacktrace: expect.any(Array)
           }
@@ -60,16 +60,16 @@ describe('graphql - error handling in development', () => {
         {
           message: 'Multiple errors occurred. Please see the details for more information.',
           extensions: {
-            code: '400',
+            code: 'MULTIPLE_ERRORS',
             details: [
               {
-                code: '400',
+                code: 'ASSERT_NOT_NULL',
                 message: 'Value is required',
                 target: 'notEmptyI',
                 stacktrace: expect.any(Array)
               },
               {
-                code: '400',
+                code: 'ASSERT_NOT_NULL',
                 message: 'Value is required',
                 target: 'notEmptyS',
                 stacktrace: expect.any(Array)
@@ -101,7 +101,7 @@ describe('graphql - error handling in development', () => {
         {
           message: 'Value 10 is not in specified range [0, 3]',
           extensions: {
-            code: '400',
+            code: 'ASSERT_RANGE',
             target: 'inRange',
             args: [10, 0, 3],
             stacktrace: expect.any(Array)
@@ -129,7 +129,7 @@ describe('graphql - error handling in development', () => {
         {
           message: 'Es sind mehrere Fehler aufgetreten.',
           extensions: {
-            code: '400',
+            code: 'MULTIPLE_ERRORS',
             details: [
               { target: 'inRange', message: 'Wert ist erforderlich' },
               { target: 'oneOfEnumValues', message: expect.stringContaining('Value "foo" is invalid') }
@@ -145,7 +145,7 @@ describe('graphql - error handling in development', () => {
 
       const log = console.warn.mock.calls[0][1]
       expect(log).toMatchObject({
-        code: '400',
+        code: 'MULTIPLE_ERRORS',
         details: [
           { target: 'inRange', message: 'Value is required' },
           { target: 'oneOfEnumValues', message: expect.stringContaining('Value "foo" is invalid') }
@@ -172,7 +172,7 @@ describe('graphql - error handling in development', () => {
         {
           message: 'Error on READ A',
           extensions: {
-            code: '500',
+            code: '',
             myProperty: 'My value A1',
             $myProperty: 'My value A2',
             my: { nested: { property: 'My value A3' } },
@@ -202,7 +202,7 @@ describe('graphql - error handling in development', () => {
         {
           message: 'Error on READ B',
           extensions: {
-            code: '500',
+            code: '',
             stacktrace: expect.any(Array)
           }
         }
@@ -228,7 +228,7 @@ describe('graphql - error handling in development', () => {
         {
           message: 'Error on READ C',
           extensions: {
-            code: '500',
+            code: '',
             stacktrace: expect.any(Array)
           }
         }
@@ -311,7 +311,7 @@ describe('graphql - error handling in development', () => {
         {
           message: 'Multiple errors occurred. Please see the details for more information.',
           extensions: {
-            code: '500',
+            code: 'MULTIPLE_ERRORS',
             details: [
               {
                 code: 'Some-Custom-Code1',
@@ -339,7 +339,7 @@ describe('graphql - error handling in development', () => {
       expect(response.data.errors[0].extensions.details[0].stacktrace[0]).not.toHaveLength(0) // Stacktrace exists and is not empty
       expect(response.data.errors[0].extensions.details[1].stacktrace[0]).not.toHaveLength(0) // Stacktrace exists and is not empty
       expect(console.error.mock.calls[0][1]).toMatchObject({
-        code: '500',
+        code: 'MULTIPLE_ERRORS',
         message: 'Multiple errors occurred. Please see the details for more information.',
         details: [
           {
@@ -417,7 +417,7 @@ describe('graphql - error handling in development', () => {
           }
         }
       ]
-      const response = await POST('/graphql', { query })
+      const response = await POST('/graphql', { query }, { headers: { 'Accept-Language': 'en' } })
       expect(response.data).toMatchObject({ errors })
       expect(response.data.errors[0].extensions.stacktrace).not.toHaveLength(0) // Stacktrace exists and is not empty
     })
@@ -440,7 +440,7 @@ describe('graphql - error handling in development', () => {
         {
           message: 'The order of 20 books exceeds the stock by 10',
           extensions: {
-            code: '500',
+            code: '',
             target: 'ORDER_EXCEEDS_STOCK',
             args: [20, 10],
             numericSeverity: 4,
@@ -448,7 +448,7 @@ describe('graphql - error handling in development', () => {
           }
         }
       ]
-      const response = await POST('/graphql', { query })
+      const response = await POST('/graphql', { query }, { headers: { 'Accept-Language': 'en' } })
       expect(response.data).toMatchObject({ errors })
       expect(response.data.errors[0].extensions.stacktrace).not.toHaveLength(0) // Stacktrace exists and is not empty
     })